+++ apps/api/v1/pages/api/slots/_get.ts
@@ -1,22 +1,14 @@
-import { createContainer } from "@evyweb/ioctopus";
 import timezone from "dayjs/plugin/timezone";
 import utc from "dayjs/plugin/utc";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import dayjs from "@calcom/dayjs";
 import { isSupportedTimeZone } from "@calcom/lib/dayjs";
-import { DI_TOKENS } from "@calcom/lib/di/tokens";
+import { getAvailableSlotsService } from "@calcom/lib/di/available-slots.container";
 import { HttpError } from "@calcom/lib/http-error";
 import { defaultResponder } from "@calcom/lib/server/defaultResponder";
-import { oooRepositoryModule } from "@calcom/lib/server/modules/ooo";
-import { scheduleRepositoryModule } from "@calcom/lib/server/modules/schedule";
-import { prismaModule } from "@calcom/prisma/prisma.module";
 import { createContext } from "@calcom/trpc/server/createContext";
 import { getScheduleSchema } from "@calcom/trpc/server/routers/viewer/slots/types";
-import {
-  availableSlotsModule,
-  type AvailableSlotsService,
-} from "@calcom/trpc/server/routers/viewer/slots/util";
 
 import { TRPCError } from "@trpc/server";
 import { getHTTPStatusCodeFromError } from "@trpc/server/http";
@@ -36,12 +28,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     const input = getScheduleSchema.parse({ usernameList: slugs, isTeamEvent: parsedIsTeamEvent, ...rest });
     const timeZoneSupported = input.timeZone ? isSupportedTimeZone(input.timeZone) : false;
 
-    const container = createContainer();
-    container.load(DI_TOKENS.PRISMA_MODULE, prismaModule);
-    container.load(DI_TOKENS.OOO_REPOSITORY_MODULE, oooRepositoryModule);
-    container.load(DI_TOKENS.SCHEDULE_REPOSITORY_MODULE, scheduleRepositoryModule);
-    container.load(DI_TOKENS.AVAILABLE_SLOTS_SERVICE_MODULE, availableSlotsModule);
-    const availableSlotsService = container.get<AvailableSlotsService>(DI_TOKENS.AVAILABLE_SLOTS_SERVICE);
+    const availableSlotsService = getAvailableSlotsService();
     const availableSlots = await availableSlotsService.getAvailableSlots({
       ctx: await createContext({ req, res }),
       input,
