+++ apps/web/pages/api/get-inbound-dynamic-variables.ts
@@ -1,4 +1,3 @@
-import { createContainer } from "@evyweb/ioctopus";
 import advancedFormat from "dayjs/plugin/advancedFormat";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { z } from "zod";
@@ -6,17 +5,9 @@ import { z } from "zod";
 import dayjs from "@calcom/dayjs";
 import { ZGetRetellLLMSchema } from "@calcom/features/ee/cal-ai-phone/zod-utils";
 import type { TGetRetellLLMSchema } from "@calcom/features/ee/cal-ai-phone/zod-utils";
-import { DI_TOKENS } from "@calcom/lib/di/tokens";
 import { fetcher } from "@calcom/lib/retellAIFetcher";
 import { defaultHandler } from "@calcom/lib/server/defaultHandler";
-import { oooRepositoryModule } from "@calcom/lib/server/modules/ooo";
-import { scheduleRepositoryModule } from "@calcom/lib/server/modules/schedule";
 import prisma from "@calcom/prisma";
-import { prismaModule } from "@calcom/prisma/prisma.module";
-import {
-  availableSlotsModule,
-  type AvailableSlotsService,
-} from "@calcom/trpc/server/routers/viewer/slots/util";
 
 dayjs.extend(advancedFormat);
 
@@ -94,12 +85,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   const startTime = now.startOf("month").toISOString();
   const endTime = now.add(2, "month").endOf("month").toISOString();
   const orgSlug = eventType?.team?.parent?.slug ?? null;
-  const container = createContainer();
-  container.load(DI_TOKENS.PRISMA_MODULE, prismaModule);
-  container.load(DI_TOKENS.OOO_REPOSITORY_MODULE, oooRepositoryModule);
-  container.load(DI_TOKENS.SCHEDULE_REPOSITORY_MODULE, scheduleRepositoryModule);
-  container.load(DI_TOKENS.AVAILABLE_SLOTS_SERVICE_MODULE, availableSlotsModule);
-  const availableSlotsService = container.get<AvailableSlotsService>(DI_TOKENS.AVAILABLE_SLOTS_SERVICE);
+  const availableSlotsService = getAvailableSlotsService();
 
   const availableSlots = await availableSlotsService.getAvailableSlots({
     input: {
